# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OMEDIA",
    "description": "Disk-first media organization API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/origins": {
      "get": {
        "tags": [
          "origins"
        ],
        "summary": "List Origins",
        "operationId": "list_origins",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OriginSummary"
                  },
                  "type": "array",
                  "title": "Response List Origins"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "origins"
        ],
        "summary": "Create Origin",
        "operationId": "create_origin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OriginDraft"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Origin"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/origins/{origin_id}": {
      "get": {
        "tags": [
          "origins"
        ],
        "summary": "Origin Detail",
        "operationId": "origin_detail",
        "parameters": [
          {
            "name": "origin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Origin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OriginSummary"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "origins"
        ],
        "summary": "Save Origin",
        "operationId": "save_origin",
        "parameters": [
          {
            "name": "origin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Origin Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OriginDraft"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Origin"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "origins"
        ],
        "summary": "Delete Origin",
        "operationId": "delete_origin",
        "parameters": [
          {
            "name": "origin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Origin Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/identify/sessions/{session_id}": {
      "post": {
        "tags": [
          "identify"
        ],
        "summary": "Identify Session",
        "operationId": "identify_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizeSession"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/identify/sessions/{session_id}/search": {
      "post": {
        "tags": [
          "identify"
        ],
        "summary": "Search Identify Session",
        "operationId": "search_identify_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentifySearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentifySearchResults"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/identify/sessions/{session_id}/candidates/{candidate_id}/override": {
      "put": {
        "tags": [
          "identify"
        ],
        "summary": "Apply Identify Candidate Override",
        "operationId": "apply_identify_candidate_override",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IdentityOverrideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizeSession"
                }
              }
            }
          },
          "400": {
            "description": "Application error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "State conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation or domain error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorDetail"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse"
      },
...
      "CandidateDecision": {
        "type": "string",
        "enum": [
          "accept",
          "ignore"
        ],
        "title": "CandidateDecision"
      },
...
      "IdentifySearchRequest": {
        "properties": {
          "source_candidate_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Candidate Id"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "default": 1
          }
        },
        "type": "object",
        "title": "IdentifySearchRequest"
      },
...
      "IdentifySearchResults": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/IdentifySearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "total_results": {
            "type": "integer",
            "title": "Total Results"
          }
        },
        "type": "object",
        "required": [
          "results",
          "page",
          "total_pages",
          "total_results"
        ],
        "title": "IdentifySearchResults"
      },
      "IdentityOverrideRequest": {
        "properties": {
          "tmdb_id": {
            "type": "integer",
            "title": "Tmdb Id"
          }
        },
        "type": "object",
        "required": [
          "tmdb_id"
        ],
        "title": "IdentityOverrideRequest"
      },
...
      "MediaType": {
        "type": "string",
        "enum": [
          "tv",
          "movie"
        ],
        "title": "MediaType"
      },
...
      "OrganizePolicy": {
        "properties": {
          "target_depot_id": {
            "type": "string",
            "title": "Target Depot Id"
          },
          "organize_rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organize Rule Id"
          }
        },
        "type": "object",
        "required": [
          "target_depot_id"
        ],
        "title": "OrganizePolicy"
      },
...
      "OrganizeSession": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "$ref": "#/components/schemas/OrganizeSessionKind"
          },
          "path": {
            "type": "string",
            "format": "path",
            "title": "Path"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "policy": {
            "$ref": "#/components/schemas/OrganizePolicy"
          },
          "state": {
            "$ref": "#/components/schemas/OrganizeSessionState"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_active_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Active At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "idle_timeout_seconds": {
            "type": "integer",
            "title": "Idle Timeout Seconds"
          },
          "review_candidates": {
            "items": {
              "$ref": "#/components/schemas/SourceCandidate"
            },
            "type": "array",
            "title": "Review Candidates"
          },
          "last_source_action_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceActionOutcome"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "path",
          "media_type",
          "policy",
          "state",
          "created_at",
          "last_active_at",
          "expires_at",
          "idle_timeout_seconds",
          "review_candidates"
        ],
        "title": "OrganizeSession"
      },
      "OrganizeSessionKind": {
        "type": "string",
        "enum": [
          "origin",
          "ad_hoc"
        ],
        "title": "OrganizeSessionKind"
      },
      "OrganizeSessionState": {
        "type": "string",
        "enum": [
          "scanning",
          "scanned",
          "identifying",
          "identified",
          "organizing",
          "done",
          "cancelled"
        ],
        "title": "OrganizeSessionState"
      },
...
      "Origin": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "format": "path",
            "title": "Path"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "trigger": {
            "$ref": "#/components/schemas/OriginTrigger"
          },
          "policy": {
            "$ref": "#/components/schemas/OrganizePolicy"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path",
          "media_type",
          "trigger",
          "policy",
          "id"
        ],
        "title": "Origin"
      },
      "OriginDraft": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "format": "path",
            "title": "Path"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "trigger": {
            "$ref": "#/components/schemas/OriginTrigger"
          },
          "policy": {
            "$ref": "#/components/schemas/OrganizePolicy"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "path",
          "media_type",
          "trigger",
          "policy"
        ],
        "title": "OriginDraft"
      },
      "OriginSummary": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "format": "path",
            "title": "Path"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "trigger": {
            "$ref": "#/components/schemas/OriginTrigger"
          },
          "policy": {
            "$ref": "#/components/schemas/OrganizePolicy"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "candidate_count": {
            "type": "integer",
            "title": "Candidate Count",
            "default": 0
          },
          "file_count": {
            "type": "integer",
            "title": "File Count",
            "default": 0
          },
          "unknown_count": {
            "type": "integer",
            "title": "Unknown Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "name",
          "path",
          "media_type",
          "trigger",
          "policy",
          "id"
        ],
        "title": "OriginSummary"
      },
      "OriginTrigger": {
        "type": "string",
        "enum": [
          "manual",
          "watch"
        ],
        "title": "OriginTrigger"
      },
...
      "SourceCandidate": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "media_type": {
            "$ref": "#/components/schemas/MediaType"
          },
          "source_root": {
            "type": "string",
            "format": "path",
            "title": "Source Root"
          },
          "source_path": {
            "type": "string",
            "format": "path",
            "title": "Source Path"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "structure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structure"
          },
          "kind": {
            "$ref": "#/components/schemas/SourceCandidateKind"
          },
          "status": {
            "$ref": "#/components/schemas/SourceCandidateStatus"
          },
          "selection_decision": {
            "$ref": "#/components/schemas/CandidateDecision"
          },
          "file_count": {
            "type": "integer",
            "title": "File Count"
          },
          "active_file_count": {
            "type": "integer",
            "title": "Active File Count"
          },
          "total_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Size"
          },
          "active_total_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Total Size"
          },
          "modified_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified Time"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/SourceFile"
            },
            "type": "array",
            "title": "Files"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CandidateMatch"
              },
              {
                "type": "null"
              }
            ]
          },
          "plan_items": {
            "items": {
              "$ref": "#/components/schemas/PlanItem"
            },
            "type": "array",
            "title": "Plan Items"
          },
          "conflict_reviews": {
            "items": {
              "$ref": "#/components/schemas/ConflictReview"
            },
            "type": "array",
            "title": "Conflict Reviews"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "id",
          "media_type",
          "source_root",
          "source_path",
          "display_name",
          "kind",
          "status",
          "selection_decision",
          "file_count",
          "active_file_count",
          "files",
          "plan_items",
          "warnings"
        ],
        "title": "SourceCandidate"
      },
...
      "SourceCandidateKind": {
        "type": "string",
        "enum": [
          "movie_file",
          "movie_folder",
          "tv_show"
        ],
        "title": "SourceCandidateKind"
      },
      "SourceCandidateStatus": {
        "type": "string",
        "enum": [
          "active",
          "deleted",
          "unavailable",
          "blocked"
        ],
        "title": "SourceCandidateStatus"
      },
      "SourceFile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "path": {
            "type": "string",
            "format": "path",
            "title": "Path"
          },
          "relative_path": {
            "type": "string",
            "format": "path",
            "title": "Relative Path"
          },
          "extension": {
            "type": "string",
            "title": "Extension"
          },
          "classification": {
            "$ref": "#/components/schemas/SourceFileClassification"
          },
          "status": {
            "$ref": "#/components/schemas/SourceFileStatus"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "modified_time": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified Time"
          },
          "planned_item_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Planned Item Ids"
          },
          "plan_status": {
            "$ref": "#/components/schemas/SourceFilePlanStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "path",
          "relative_path",
          "extension",
          "classification",
          "status",
          "planned_item_ids",
          "plan_status"
        ],
        "title": "SourceFile"
      },
      "SourceFileClassification": {
        "type": "string",
        "enum": [
          "video",
          "subtitle",
          "generic_sidecar",
          "unsupported",
          "deleted",
          "missing",
          "blocked"
        ],
        "title": "SourceFileClassification"
      },
...
    }
  }
}